[tvOS] Add minimum set of required backports + tvos-arm64-simulator build config (#7107)
This commit includes a series of backports from upstream Chromium that make
it possible to use Chromium's toolchain (including the compiler) to target a
tvOS build. This is in no way a complete list of backports, but rather the
minimum amount of changes we need to be able to 1) specify that a Cobalt
build should target tvOS in GN and in the Cobalt scripts 2) pass the right
arguments to the toolchain 3) detect if we are doing a tvOS build from C++.

The last commit in the series then adds the tvos-arm64-simulator Cobalt
build configuration, so that one can then do

    cobalt/build/gn.py -p tvos-arm64-simulator

and build the `starboard` target (which at the moment is just a handful of
files designed to test that the build is working).

Note: as in the upstream build, `.gclient` needs `target_os=['ios']` so that
`gclient sync` fetches all required iOS/tvOS dependencies.

The focus of this commit is on adding all the required scaffolding: setting
the appropriate GN variables to the right values, adding tvOS configurations
to `cobalt/build` and `starboard/build`, and adding the minimum amount of
C++ and GN configuration files under `starboard/tvos` so that we can
hopefully add tvOS to CI.

The dependency on the cobalt targets in `//BUILD.gn` has been temporarily
disabled for tvOS builds because they depend on further changes to build
cobalt shell as a bundle, which will come later.

Bug: 432823234
Bug: 432824509
Bug: 433180638
Bug: 442923940

original-hexsha: 7b849b9126d9d269699b4b0c691219dca2215587